--- frontend/src/PlanInviteVerification.tsx
@@ -1,62 +1,0 @@
-import axios from 'axios'
-import { useEffect, useState } from 'react'
-import { useParams } from 'react-router-dom'
-
-export const PlanInviteVerification = (): JSX.Element => {
-  const { id: planInviteID } = useParams()
-  const [isWrongEmail, setIsWrongEmail] = useState<boolean>(false)
-
-  useEffect(() => {
-    const sendTokenToBackend = async () => {
-      try {
-        const response = await axios.get(`/verify-plan-invite/${planInviteID}`, { withCredentials: true })
-        return response
-      } catch (error) {
-        console.error(error)
-        return error
-      }
-    }
-
-    const ONE_SECONDS = 1000;
-
-    const delayLoading = async () => {
-      await new Promise((resolve) => setTimeout(resolve, ONE_SECONDS));
-      sendTokenToBackend()
-        .then((response: any) => {
-          switch (response.data.reason) {
-            case 'NOT_LOGGED_IN':
-              window.location.href = '/plan-invite-login-request'
-              break;
-            case 'USER_ALREADY_A_PLAN_MEMBER':
-            case 'LOGGED_IN_AND_EMAIL_MATCH':
-              const planID = response.data.planID
-              window.location.href = `/plan/${planID}`
-              break;
-            case 'LOGGED_IN_AND_EMAIL_DONT_MATCH':
-              setIsWrongEmail(true)
-              break;
-          }
-        })
-        .catch((response) => {
-          console.error(`unable to verify token ${response}`)
-
-        })
-    };
-
-    void delayLoading();
-
-  }, [])
-
-  const style = {
-    padding: '20px',
-    marginTop: '40px',
-    fontSize: '40px'
-  }
-  return (
-    <div>
-      {isWrongEmail
-        ? <h1 style={style}> Unauthorized acess: Please log into the account with the correct email address </h1>
-        : <h1 style={style}>Loading...</h1>}
-    </div>
-  )
-}
