+++ frontend/src/components/PlanInviteLogInRequest.tsx
@@ -1,12 +1,16 @@
+/*
+  This component is to display a request message for the user to log in or sign up before accepting a plan invitation
+*/
 export const PlanInviteLogInRequest = () => {
   const style = {
     padding: '20px',
     marginTop: '40px',
-    fontSize: '40px'
+    fontSize: '40px',
+    marginLeft: '400px'
   }
   return (
     <>
-      <h1 style={style}>Please login or sign up before accepting a plan invitation</h1>
+      <h3 style={style}>Please login or sign up before accepting a plan invitation</h3>
     </>
   )
 }
